requirements.txt
Django~=4.2.20
djangorestframework~=3.16.0
psycopg2~=2.9.10